[#28] refactor : 피드백 반영
1. 파일 위치 변경
