+++ src/main/java/com/btb/chalKak/ChalKakApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class ChalKakApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ChalKakApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ChalKakApplication.class, args);
+    }
 
 }
